A bunch of minor fixes.  Some declarations not at beginning of block moved there for MSVC.  Changed completion flag to use bits rather than bytes in possible preparation for answer_complete flag.  Some format fixes.
git-svn-id: svn+ssh://svn.code.sf.net/p/xsb/src/trunk@8401 756869e0-9e47-4bc5-b0a8-847305dc5e1c

+++ XSB/emu/pathname_xsb.c
@@ -546,7 +546,7 @@ void fix_cygwin_pathname(char *filename)
   if (strlen(filename) < MAXPATHLEN - 9    // /cygdrive = 9 chars
       && filename[1] == ':'
       && filename[2] == '/'
-      && isalpha(filename[0]))
+      && isalpha(((int)filename[0])))
     {
       strcpy(tmp,"/cygdrive/");
       tmp[10] = filename[0];
